+++ PluginsAndFeatures/azure-toolkit-for-rider/azure-intellij-plugin-cloud-shell/src/main/kotlin/com/microsoft/azure/toolkit/intellij/cloudshell/terminal/AzureCloudTerminalProcess.kt
@@ -4,7 +4,6 @@
 
 package com.microsoft.azure.toolkit.intellij.cloudshell.terminal
 
-import com.microsoft.azure.toolkit.intellij.cloudshell.rest.CloudConsoleTerminalWebSocket
 import org.jetbrains.plugins.terminal.cloud.CloudTerminalProcess
 
 class AzureCloudTerminalProcess(private val socketClient: CloudConsoleTerminalWebSocket) :
