+++ PluginsAndFeatures/azure-toolkit-for-rider/azure-intellij-plugin-cloud-shell/src/main/kotlin/com/microsoft/azure/toolkit/intellij/cloudshell/terminal/CloudConsoleTerminalWebSocket.kt
@@ -2,7 +2,7 @@
  * Copyright 2018-2024 JetBrains s.r.o. and contributors. Use of this source code is governed by the MIT license.
  */
 
-package com.microsoft.azure.toolkit.intellij.cloudshell.rest
+package com.microsoft.azure.toolkit.intellij.cloudshell.terminal
 
 import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.diagnostic.logger
